requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
cffi==1.16.0
click==8.1.7
cryptography==43.0.0
ecdsa==0.19.0
fastapi==0.112.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
jwt==1.3.1
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
redis==5.0.8
rsa==4.9
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
